Group name instead group id

+++ app/assets/javascripts/backbone/apps/student/list/list_view.js.coffee
@@ -10,19 +10,22 @@ App.module "StudentApp.List", (List, App, Backbone, Marionette, $, _) ->
     tagName: 'tr'
 
     editStudent: ->
+      console.log @model.attributes
       $("#input-last-name-edit").val(@model.attributes.last_name)
       $("#input-name-edit").val(@model.attributes.name)
-      $("#input-group-edit").val(@model.attributes.group)
+#      $("#input-group-edit").val(@model.attributes.group_id)
+      $("#select-group-edit option:selected").text(@model.attributes.group_name)
       $("#input-email-edit").val(@model.attributes.email)
       $("#input-date-of-birth-edit").val(@model.attributes.date_of_birth)
       $("#input-id-edit").val(@model.attributes.id)
 
     deleteStudent: ->
-     console.log @model.destroy
-       url: Routes.delete_student_path({ format: 'json', id: @model.id }),
-       type: "post",
-       success: =>
-         console.log "destroy"
+      if confirm("Удалить запись?") == true
+        console.log @model.destroy
+          url: Routes.delete_student_path({ format: 'json', id: @model.id }),
+          type: "post",
+          success: =>
+            console.log "Удалена"
 
   class List.StudentsCompositeView extends App.Base.CompositeView
 
@@ -32,10 +35,12 @@ App.module "StudentApp.List", (List, App, Backbone, Marionette, $, _) ->
 
   class List.LayoutView extends App.Base.LayoutView
     initialize: ->
+      @selectGroup()
 
     events:
       'submit #form-create-new-student': 'submitForm'
       'submit #form-edit-student': 'editStudent'
+      'click #myModal': 'emptyValue'
 
     submitForm: (e) ->
       e.preventDefault()
@@ -45,10 +50,15 @@ App.module "StudentApp.List", (List, App, Backbone, Marionette, $, _) ->
         "id": @$el.find('#input-id').val()
         "last_name": @$el.find('#input-last-name').val()
         "name": @$el.find('#input-name').val()
-        "group": @$el.find('#input-group').val()
+        "group_id": @$el.find('#select-group option:selected').val()
         "email": @$el.find('#input-email').val()
         "date_of_birth": @$el.find('#input-date-of-birth').val()
       @trigger 'submit:button:click'
+      App.vent.on "empty:value:student", =>
+        @$el.find('#input-last-name').val('')
+        @$el.find('#input-name').val('')
+        @$el.find('#input-email').val('')
+        @$el.find('#input-date-of-birth').val('')
 
     editStudent: (e)->
       e.preventDefault()
@@ -56,10 +66,17 @@ App.module "StudentApp.List", (List, App, Backbone, Marionette, $, _) ->
         "id": @$el.find('#input-id-edit').val()
         "last_name": @$el.find('#input-last-name-edit').val()
         "name": @$el.find('#input-name-edit').val()
-        "group": @$el.find('#input-group-edit').val()
+        "group_id": @$el.find('#select-group-edit option:selected').val()
         "email": @$el.find('#input-email-edit').val()
         "date_of_birth": @$el.find('#input-date-of-birth-edit').val()
       @trigger 'submit:edit:button:click'
+
+    selectGroup: ->
+      @$el.find('#select-group').change =>
+        console.log "Select Group"
+
+    emptyValue: ->
+      console.log "ClickModal"
